Refactor: move CFL methods to seperate header file
